Move AutomatedSessionRemindersTest to new E2E test package
