+++ src/e2e/java/teammates/e2e/cases/e2e/AutomatedSessionRemindersTest.java
@@ -1,15 +1,14 @@
-package teammates.test.cases.browsertests;
+package teammates.e2e.cases.e2e;
 
 import org.testng.annotations.Test;
 
 import teammates.common.datatransfer.attributes.FeedbackSessionAttributes;
 import teammates.common.util.AppUrl;
-import teammates.common.util.Config;
 import teammates.common.util.Const;
 import teammates.common.util.TimeHelper;
 import teammates.common.util.TimeHelperExtension;
-import teammates.e2e.util.Priority;
-import teammates.test.pageobjects.GenericAppPage;
+import teammates.e2e.pageobjects.GenericAppPage;
+import teammates.e2e.util.TestProperties;
 
 /**
  * SUT: {@link Const.CronJobURIs#AUTOMATED_FEEDBACK_OPENING_REMINDERS},
@@ -17,27 +16,25 @@ import teammates.test.pageobjects.GenericAppPage;
  *      {@link Const.CronJobURIs#AUTOMATED_FEEDBACK_CLOSED_REMINDERS},
  *      {@link Const.CronJobURIs#AUTOMATED_FEEDBACK_PUBLISHED_REMINDERS}.
  */
-@Priority(5)
-public class AutomatedSessionRemindersTest extends BaseLegacyUiTestCase {
+public class AutomatedSessionRemindersTest extends BaseE2ETestCase {
 
     @Override
     protected void prepareTestData() {
         testData = loadDataBundle("/AutomatedSessionRemindersTest.json");
 
-        /*
-         * In this test, we set the email address of the accounts to be the same as the
-         * support email address. When running the test against a production server,
-         * email alerts will be sent to the specified support email address.
-         * The tester should manually check the email box after running the test suite.
-         */
+        // When running the test against a production server, email alerts will be sent
+        // to the specified email address
+        // The tester should manually check the email box after running the test suite
+        // TODO check if we can automate this checking process
 
-        testData.accounts.get("instructorWithEvals").email = Config.SUPPORT_EMAIL;
-        testData.instructors.get("AutSessRem.instructor").email = Config.SUPPORT_EMAIL;
-        testData.students.get("alice.tmms@AutSessRem.course").email = Config.SUPPORT_EMAIL;
-        testData.feedbackSessions.get("closedSession").setCreatorEmail(Config.SUPPORT_EMAIL);
-        testData.feedbackSessions.get("closingSession").setCreatorEmail(Config.SUPPORT_EMAIL);
-        testData.feedbackSessions.get("openingSession").setCreatorEmail(Config.SUPPORT_EMAIL);
-        testData.feedbackSessions.get("publishedSession").setCreatorEmail(Config.SUPPORT_EMAIL);
+        String student1Email = TestProperties.TEST_STUDENT1_ACCOUNT + "@gmail.com";
+        testData.accounts.get("instructorWithEvals").email = student1Email;
+        testData.instructors.get("AutSessRem.instructor").email = student1Email;
+        testData.students.get("alice.tmms@AutSessRem.course").email = student1Email;
+        testData.feedbackSessions.get("closedSession").setCreatorEmail(student1Email);
+        testData.feedbackSessions.get("closingSession").setCreatorEmail(student1Email);
+        testData.feedbackSessions.get("openingSession").setCreatorEmail(student1Email);
+        testData.feedbackSessions.get("publishedSession").setCreatorEmail(student1Email);
 
         // Set closing time of one feedback session to tomorrow
         FeedbackSessionAttributes closingFeedbackSession = testData.feedbackSessions.get("closingSession");
@@ -51,33 +48,37 @@ public class AutomatedSessionRemindersTest extends BaseLegacyUiTestCase {
         FeedbackSessionAttributes openingFeedbackSession = testData.feedbackSessions.get("openingSession");
         openingFeedbackSession.setStartTime(TimeHelper.getInstantDaysOffsetFromNow(-1));
 
-        //Published time for one feedback session already set to some time in the past.
+        // Published time for one feedback session already set to some time in the past.
 
         removeAndRestoreDataBundle(testData);
     }
 
+    // In all the tests, we need to explicitly log in as admin
+    // because the cron job URLs are additionally protected by admin constraint in web.xml
+    // and adding backdoor key is not sufficient to bypass it
+
     @Test
     public void testFeedbackSessionOpeningReminders() {
         AppUrl openingRemindersUrl = createUrl(Const.CronJobURIs.AUTOMATED_FEEDBACK_OPENING_REMINDERS);
-        loginAdminToPageOld(openingRemindersUrl, GenericAppPage.class);
+        loginAdminToPage(openingRemindersUrl, GenericAppPage.class);
     }
 
     @Test
     public void testFeedbackSessionClosingReminders() {
         AppUrl closingRemindersUrl = createUrl(Const.CronJobURIs.AUTOMATED_FEEDBACK_CLOSING_REMINDERS);
-        loginAdminToPageOld(closingRemindersUrl, GenericAppPage.class);
+        loginAdminToPage(closingRemindersUrl, GenericAppPage.class);
     }
 
     @Test
     public void testFeedbackSessionClosedReminders() {
         AppUrl closedRemindersUrl = createUrl(Const.CronJobURIs.AUTOMATED_FEEDBACK_CLOSED_REMINDERS);
-        loginAdminToPageOld(closedRemindersUrl, GenericAppPage.class);
+        loginAdminToPage(closedRemindersUrl, GenericAppPage.class);
     }
 
     @Test
     public void testFeedbackSessionPublishedReminders() {
         AppUrl publishedRemindersUrl = createUrl(Const.CronJobURIs.AUTOMATED_FEEDBACK_PUBLISHED_REMINDERS);
-        loginAdminToPageOld(publishedRemindersUrl, GenericAppPage.class);
+        loginAdminToPage(publishedRemindersUrl, GenericAppPage.class);
     }
 
 }
